requirements.txt
attrs==17.4.0
backcall==0.1.0
decorator==4.3.0
Django==2.0.4
django-rest-framework==0.1.0
django-test-plus==1.0.22
djangorestframework==3.8.2
factory-boy==2.10.0
Faker==0.8.13
ipdb==0.11
ipython==6.3.1
ipython-genutils==0.2.0
jedi==0.12.0
more-itertools==4.1.0
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
py==1.5.3
Pygments==2.2.0
python-dateutil==2.7.2
pytz==2018.4
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
wcwidth==0.1.7
